% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{mybin}
\alias{mybin}
\title{Binomial Function}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{Number Of Iterations}

\item{n}{Sample Size}

\item{p}{Probability Of Success}
}
\value{
Binomial Distribution Plot
}
\description{
Binomial Function
}
\examples{
\dontrun{mybin(iter=100, n=10, p=0.7)}

}
