% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{sample}

\item{fun}{fun}

\item{alpha}{1- percent confidence}

\item{cx}{Size}

\item{...}{additional arguments}
}
\value{
Plot
}
\description{
myboot2
}
\examples{
\dontrun{myboot2(iter=10000,sample=x,fun="mean",alpha=0.05,cx=1.5,...)}

}
