% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLTuniform.R
\name{mycltu}
\alias{mycltu}
\title{CLTuniform}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Number of Iterations}

\item{a}{a}

\item{b}{b}
}
\value{
Plot
}
\description{
CLTuniform
}
\examples{
\dontrun{mycltu(n=5,iter=10000)}

}
