% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{My Example Function}
\usage{
myfun(x)
}
\arguments{
\item{x}{Numerical Value}
}
\value{
Plot Of x Vs y, Where y Is X^2
}
\description{
My Example Function
}
\examples{
\dontrun{myfun(x = 1:10)}

}
