% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Probability Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the deviation}

\item{a}{the value X is <= to}
}
\value{
Curve And Probability
}
\description{
Normal Probability Curve
}
\examples{
#' \dontrun{myncurve(mu=10,sigma=5,a=6)}

}
