% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysd.R
\name{mysd}
\alias{mysd}
\title{My Standard Deviation Function}
\usage{
mysd(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numerical Vector}

\item{na.rm}{Logical Indicating Whether To Remove NA Values Before Computation}
}
\value{
Standard Deviation Of The Numeric Vector
}
\description{
My Standard Deviation Function
}
\examples{
\dontrun{1}

}
