% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter Histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{X Variable Vector}

\item{y}{Y Variable Vector}

\item{xlab}{x Label}

\item{ylab}{y Label}
}
\value{
Plot
}
\description{
Scatter Histogram
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}

}
