% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum likelihood function (Normal)}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{x}

\item{mu}{mean}

\item{sig}{variance}

\item{...}{addition arguments}
}
\value{
Plot
}
\description{
Maximum likelihood function (Normal)
}
\examples{
\dontrun{mymlnorm(x=c(10,12,13,15,12,11,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)}
}
